fix(core): 修正版本号类型支持大数值
- 将 Beta 版本号类型由 int? 修改为 long?，支持更大数值如时间戳
- 更新构造函数及解析逻辑以使用 long 类型 BetaNo
- PowerShell 启动脚本中添加 --show 参数以确保界面显示
- .gitignore 文件新增发布测试目录及版本检查脚本忽略规则

+++ src/MacroDeck/DataTypes/Core/Version.cs
@@ -7,9 +7,10 @@ public partial struct Version
     public int Major { get; set; }
     public int Minor { get; set; }
     public int Patch { get; set; }
-    public int? BetaNo { get; set; }
+    /// <summary>Beta 版本号，使用 long 类型以支持时间戳等大数值构建号。</summary>
+    public long? BetaNo { get; set; }
 
-    public Version(int major, int minor, int patch, int? betaNo = null)
+    public Version(int major, int minor, int patch, long? betaNo = null)
     {
         Major = major;
         Minor = minor;
@@ -59,10 +60,10 @@ public partial struct Version
         var minor = int.Parse(match.Groups["minor"].Value);
         var patch = int.Parse(match.Groups["patch"].Value);
 
-        int? previewNo = null;
+        long? previewNo = null;
         if (match.Groups["beta"].Success)
         {
-            previewNo = int.Parse(match.Groups["beta"].Value);
+            previewNo = long.Parse(match.Groups["beta"].Value);
         }
 
         return new Version(major, minor, patch, previewNo);
